+++ CHANGES.rst
@@ -8,7 +8,7 @@ History
   [kakshay21]
 
 - Fix #33 ConfigParser.NoOptionError
-[kakshay21]
+  [kakshay21]
 
 
 0.2.1 (2018-07-09)